requirements.txt
Python version: 3.12.0
PyTorch version: 2.1.1
h5py==3.1.0
numpy==1.19.5
Pillow==8.2.0
tqdm
